feat(ga): move files to a new folder
